% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_RVM.R
\docType{data}
\name{RVM}
\alias{RVM}
\title{RVIneMatrix object from VineCopula package}
\format{a \code{RVineMatrix}
\describe{
\item{RVM}{RVineMatrix}
}}
\usage{
data(RVM)
}
\description{
A RVineMatrix that was estimated using the VineCopula package on 23 
companies in the Top40 index of South Africa
}
\keyword{datasets}

